with float -> double

+++ src/main/scala/hackrf/HackRFSweepDataCallback.java
@@ -7,5 +7,5 @@ public interface HackRFSweepDataCallback {
      * @param frequencyStart array of fft bin's start frequencies, null if no data
      * @param signalPowerdBm array of fft bin's power in dB, null if no data
      */
-    void newSpectrumData(double[] frequencyStart, float[] signalPowerdBm);
+    void newSpectrumData(double[] frequencyStart, double[] signalPowerdBm);
 }
